Feature EPL-572: Upgrade Core Version with Gradle Task
This commit creates the Gradle task upgradeCoreVersion to upgrade the Core version of a project based on the type property passed to the task. If type is release, the task updates the version to the next release, and if type is hotfix, it updates the version to the next hotfix. The task also updates the version in two AD_MODULE.xml files, and the build.gradle file with the new values. If the format of the AD_MODULE.xml files was changed, the task throws a GradleException.