Issue #481: Support for scientific notation in bigdecimal conversion
ETP-377: The function in charge of converting the representation of the number entered by the user to a secure format has been modified. This representation is in string format and needs to be corrected in some cases. However, the implementation of the function did not support scientific notation, so it has been modified to support it.