Issue #516: Fix for OBPickAndExecuteGrid
ETP-691: When a filter is applied to an OBPickAndExecuteGrid and then a record is selected, when scrolling, the dataSource.transformRequest function checks the selected records and if they are not within the request criteria, it tries to add them. Here, information about the grid parameters is lost, so corrections have been made to recover them.